Update patch set 1
Patch Set 1:

Not sure if this change was part of the 17th nightly. The file still does not have +x mode bit set. I am not sure if the +x mode bit set in git will remain intact after being copied to phone during flash. If the mode bits were carried from git to phone, we would not have to fix the permissions in the updater-script. Looks like we still need an entry in the updater-script.

Patch-set: 1