Update patch set 1
Patch Set 1:

Should this also be a rename and not an addition? I kind of figured in general that adding it was better than renaming so that other people could more easily use our stuff for just n8013, but idk. The one tricky thing is that it might make sense for there to still be n8013/n8010 device trees for TWRP purposes so that devices are identified properly when installing stuff, but I guess the userbase is so small it's not that big of a deal anyway. It's probably fine to have an n801x recovery that has ro.product.device set to GT-N8013 so that everyone can install N8013 ROMs (I don't think there even really are ROMs that target N8010 specifically).

Patch-set: 1