Update patch set 2
Patch Set 2:

Using the recovery that is built when building OMNI results in TWRP without an adjustable brightness.  Is there a flag required to enable that?

The storage options are Micro SDCard as well as USB-OTG. There is no option for internal memory.

I guess I have built this incorrectly??

Patch-set: 2