Update patch set 2
Patch Set 2: Code-Review+1

Are you sure you have this patch (either PS1 or PS2)? I think as long as the path is defined it gives you an adjustable brightness, but could you confirm that the path I used is present on your device?

My experience with a recovery built using PS1 was that the brightness slider was there, but changing it did nothing, and the path in PS1 was not present on my device, so I assume the slider will be there regardless of whether the path exists.

Patch-set: 2
Reviewer: Gerrit User 1000183 <1000183@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=+1