Update patch set 2
Patch Set 2:

> Look at #11515, this file seems to be used by n80xx. Maybe #11535
 > would be better than this change.

As far as I can tell this only relevant to diplay issues in LPM charging.
n800x uses it'S own solution for fixing lpm charging and already has an updated grpahics.c in it's device tree.
On i9300 there is no need for this to have functioning lpm-charging (maybe n8000 neither, but marco hast to look into that). Either way I think this is something that should be device specific if needed and not in the common repo.

Patch-set: 2