Update patch set 1
Patch Set 1: Code-Review-2

> I think we don't have any flag to check it device is Wifi only. But
 > they didn't inherit this
 > 
 > $(call inherit-product, $(SRC_TARGET_DIR)/product/full_base_telephony.mk)

Sadly this gets included after common.mk, otherwise we could check for rild or mms in PRODUCT_PACKAGES. If noone else has a good idea how to check for it I don't think we can conditionally prevent the addition of the prop and in turn the pollution of non phone devices. So I think it's best to keep this in the device specific makefile.

Patch-set: 1
Reviewer: Gerrit User 1001140 <1001140@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=-2