Update patch set 1
Patch Set 1: Code-Review-1

I am NOT using this. It was a mismerge by me. This appears to have zero effect so I removed it and when I picked it from myself that got thrown in. Thanks for separating this from the actual fix and it was my derp. Fell asleep before I could push ps2 and remove

Patch-set: 1